#28 Move invites to new dashboard section
